python-agent/requirements.txt
livekit>=1.0.0
livekit-agents>=1.2.0
livekit-plugins-openai>=1.2.0
livekit-plugins-bey>=1.2.0
livekit-plugins-deepgram>=1.2.0
livekit-plugins-silero>=1.2.0
python-dotenv>=1.0.0
aiohttp>=3.9.0